Hard code EFS mount to early boot.
This is unfortunate.  Because of private bind propagation docker daemon
does not see the /efs NFS mount that is made after it's launched.  I've
not found an elegant way around this yet and I can't seem to find a way
to modify the docker daemon execution to first use shared propagation.

For now I'm going to hard code the EFS mount to a specific volume.  This
is a HACK.  I apologize in advance.